[OAP-Cache] Decouple DAGScheduler.scala (#1305)
In OAP, we change DAGScheduler.scala to make sure DAGScheduler.getPreferredLocsInternal will return the cache location as the prefered location. But we can also achieve the same functionality as long as we return the host address as the form like "executor_[hostname]_[executorid]" which will be recognized as a ExecutorCacheTaskLocation by Spark automatically. Search TaskLocation code for more details.